Pull out faucet class
Closes #65
