+++ packages/faucet/package.json
@@ -40,6 +40,7 @@
     "@iov/crypto": "^2.0.0-alpha.7",
     "@iov/encoding": "^2.0.0-alpha.7",
     "@iov/keycontrol": "^2.0.0-alpha.7",
+    "@iov/utils": "^2.0.0-alpha.7",
     "@koa/cors": "^3.0.0",
     "axios": "^0.19.0",
     "fast-deep-equal": "^3.1.1",
